🌠

+++ ast.hpp
@@ -6,10 +6,10 @@ namespace mavka::ast {
     class ASTNode {
     public:
         std::string TYPE = "ASTNode";
-        int start_line = 0;
-        int start_column = 0;
-        int end_line = 0;
-        int end_column = 0;
+        size_t start_line = 0;
+        size_t start_column = 0;
+        size_t end_line = 0;
+        size_t end_column = 0;
 
         virtual ~ASTNode() = default;
     };
